backend/requirements.txt
# 核心框架
setuptools==68.0.0  # 必须使用新版本
Flask==3.0.2
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.1
Flask-Talisman==1.1.0
python-dotenv==1.0.1
gunicorn==21.2.0  # 添加 gunicorn 作为 WSGI 服务器
prometheus-flask-exporter==0.23.0
healthcheck==1.3.3

# 数据库驱动
pymysql==1.1.0
redis==5.0.1
SQLAlchemy==2.0.27
SQLAlchemy-Utils==0.41.1

# 爬虫相关
playwright==1.41.2  # 替代pyppeteer
requests==2.31.0
beautifulsoup4==4.12.3
tldextract==5.1.1
fake-useragent==1.4.0
pyotp==2.9.0
user-agents==2.2.0

# 异步任务
celery==5.3.6  # 必须升级
eventlet==0.35.1
flower==2.0.1

# 监控和日志
sentry-sdk==1.40.5
python-json-logger==2.0.7
prometheus_client==0.20.0

# 安全
cryptography==42.0.5
argon2-cffi==23.1.0
certifi==2024.2.2
PyJWT==2.8.0

# 测试和开发工具
pytest==8.0.2
pytest-cov==4.1.0
pytest-flask==1.3.0
black==24.2.0
flake8==7.0.0
mypy==1.8.0